requirements.txt
python-dotenv
flask-cors
flask
duffel-api 
langchain 
langchain-core 
pydantic 
openai 
langchain_openai
langchain-cli
langsmith